Flytt møtebehovkomponenter i egen undermappe, Legg komponentene fra SykmeldingMotebehovVisning i egne filer
